% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_historical_data}
\alias{get_historical_data}
\title{Get historical data}
\usage{
get_historical_data(id, from, to)
}
\arguments{
\item{id}{Coin ID as provided by CoinMarketCap.com. Use the `get_ticker()` function to get a full list of availble IDs and coin names.}

\item{from}{Start date in date format}

\item{to}{End date in date format}
}
\value{
A tibble with historical data (date, open, high, low, close, volume, market cap) for a specific coin
}
\description{
This function scraps historical data from CoinMarketCap.com
}
