% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_ticker}
\alias{get_ticker}
\title{Get tickers}
\usage{
get_ticker()
}
\value{
A tibble with all coin IDs and names available on CoinMarketCap.com
}
\description{
Get all coin IDs and names available on CoinMarketCap.com, including current data
}
